Move language metadata stuff into new package
