+++ core/spoofax.compiler/src/main/resources/mb/spoofax/compiler/spoofaxcore/adapter_project/Component.java.mustache
@@ -6,7 +6,7 @@ import mb.spoofax.core.language.LanguageScope;
 import mb.spoofax.core.platform.PlatformComponent;
 {{#isMultiLang}}
 import mb.statix.multilang.MultiLangComponent;
-import mb.statix.multilang.LanguageMetadataProvider;
+import mb.statix.multilang.metadata.LanguageMetadataProvider;
 {{/isMultiLang}}
 
 @LanguageScope
