+++ core/spoofax.compiler/src/main/resources/mb/spoofax/compiler/spoofaxcore/adapter_project/Module.java.mustache
@@ -32,13 +32,13 @@ import mb.spoofax.core.platform.Platform;
 import mb.spoofax.core.pie.PieProvider;
 import mb.spoofax.core.pie.DefaultPieProvider;
 {{#isMultiLang}}
-import mb.statix.multilang.AnalysisContextService;
-import mb.statix.multilang.ContextDataManager;
-import mb.statix.multilang.ContextPieManager;
-import mb.statix.multilang.ImmutableLanguageMetadata;
-import mb.statix.multilang.LanguageId;
-import mb.statix.multilang.LanguageMetadata;
-import mb.statix.multilang.LanguageMetadataManager;
+import mb.statix.multilang.metadata.AnalysisContextService;
+import mb.statix.multilang.metadata.ContextDataManager;
+import mb.statix.multilang.metadata.ContextPieManager;
+import mb.statix.multilang.metadata.ImmutableLanguageMetadata;
+import mb.statix.multilang.metadata.LanguageId;
+import mb.statix.multilang.metadata.LanguageMetadata;
+import mb.statix.multilang.metadata.LanguageMetadataManager;
 import mb.statix.multilang.MultiLang;
 import mb.statix.multilang.SharedPieProvider;
 import mb.statix.multilang.pie.SmlSolveProject;
