+++ core/spoofax.compiler/src/main/resources/mb/spoofax/compiler/spoofaxcore/multilang_analyzer/SmlCheckTaskDef.java.mustache
@@ -7,8 +7,8 @@ import mb.pie.api.ResourceStringSupplier;
 import mb.pie.api.Supplier;
 import mb.resource.ResourceKey;
 import mb.spoofax.core.language.LanguageScope;
-import mb.statix.multilang.LanguageId;
-import mb.statix.multilang.LanguageMetadataManager;
+import mb.statix.multilang.metadata.LanguageId;
+import mb.statix.multilang.metadata.LanguageMetadataManager;
 import mb.statix.multilang.pie.SmlSolveProject;
 import mb.statix.multilang.pie.SmlCheckTaskDef;
 import mb.statix.multilang.pie.config.SmlBuildContextConfiguration;
