+++ core/statix.multilang.eclipse/src/main/java/mb/statix/multilang/eclipse/LanguageMetadataProvider.java
@@ -1,8 +1,8 @@
 package mb.statix.multilang.eclipse;
 
-import mb.statix.multilang.ContextId;
-import mb.statix.multilang.LanguageId;
-import mb.statix.multilang.LanguageMetadata;
+import mb.statix.multilang.metadata.ContextId;
+import mb.statix.multilang.metadata.LanguageId;
+import mb.statix.multilang.metadata.LanguageMetadata;
 
 import java.util.Map;
 import java.util.function.Function;
