+++ core/statix.multilang.eclipse/src/main/java/mb/statix/multilang/eclipse/MultiLangPlugin.java
@@ -3,8 +3,8 @@ package mb.statix.multilang.eclipse;
 import mb.log.api.Logger;
 import mb.spoofax.eclipse.SpoofaxEclipseComponent;
 import mb.spoofax.eclipse.SpoofaxPlugin;
-import mb.statix.multilang.AnalysisContextService;
-import mb.statix.multilang.ImmutableAnalysisContextService;
+import mb.statix.multilang.metadata.AnalysisContextService;
+import mb.statix.multilang.metadata.ImmutableAnalysisContextService;
 import mb.statix.multilang.MultiLangModule;
 import org.checkerframework.checker.nullness.qual.NonNull;
 import org.checkerframework.checker.nullness.qual.Nullable;
