+++ core/statix.multilang/src/main/java/mb/statix/multilang/MultiLangComponent.java
@@ -2,6 +2,7 @@ package mb.statix.multilang;
 
 import dagger.Component;
 import mb.spoofax.core.platform.PlatformComponent;
+import mb.statix.multilang.metadata.AnalysisContextService;
 import mb.statix.multilang.pie.SmlBuildSpec;
 import mb.statix.multilang.pie.SmlInstantiateGlobalScope;
 import mb.statix.multilang.pie.SmlPartialSolveFile;
@@ -16,6 +17,7 @@ import mb.statix.multilang.pie.config.SmlReadConfigYaml;
     dependencies = PlatformComponent.class
 )
 public interface MultiLangComponent {
+
     @MultiLang AnalysisContextService getAnalysisContextService();
 
     @MultiLang SmlSolveProject getAnalyzeProject();
