+++ core/statix.multilang/src/main/java/mb/statix/multilang/metadata/ContextDataManager.java
@@ -1,4 +1,4 @@
-package mb.statix.multilang;
+package mb.statix.multilang.metadata;
 
 import java.util.Set;
