+++ core/statix.multilang/src/main/java/mb/statix/multilang/metadata/ContextPieManager.java
@@ -1,6 +1,7 @@
-package mb.statix.multilang;
+package mb.statix.multilang.metadata;
 
 import mb.pie.api.Pie;
+import mb.statix.multilang.MultiLangAnalysisException;
 
 import java.util.Collection;
