+++ core/statix.multilang/src/main/java/mb/statix/multilang/metadata/LanguageMetadata.java
@@ -1,4 +1,4 @@
-package mb.statix.multilang;
+package mb.statix.multilang.metadata;
 
 import mb.common.result.Result;
 import mb.pie.api.Function;
