+++ core/statix.multilang/src/main/java/mb/statix/multilang/metadata/LanguageMetadataManager.java
@@ -1,6 +1,7 @@
-package mb.statix.multilang;
+package mb.statix.multilang.metadata;
 
 import mb.common.result.Result;
+import mb.statix.multilang.MultiLangAnalysisException;
 
 public interface LanguageMetadataManager {
     Result<LanguageMetadata, MultiLangAnalysisException> getLanguageMetadataResult(LanguageId languageId);
