+++ core/statix.multilang/src/main/java/mb/statix/multilang/pie/FileKey.java
@@ -1,10 +1,10 @@
-package mb.statix.multilang;
+package mb.statix.multilang.pie;
 
 import mb.resource.ResourceKey;
+import mb.statix.multilang.metadata.LanguageId;
 import org.immutables.value.Value;
 
 import java.io.Serializable;
-import java.util.Objects;
 
 @Value.Immutable
 public interface FileKey extends Serializable {
