+++ core/statix.multilang/src/main/java/mb/statix/multilang/pie/FileResult.java
@@ -1,4 +1,4 @@
-package mb.statix.multilang;
+package mb.statix.multilang.pie;
 
 import mb.statix.solver.persistent.SolverResult;
 import org.immutables.value.Value;
