+++ core/statix.multilang/src/main/java/mb/statix/multilang/pie/SmlBuildSpec.java
@@ -8,9 +8,9 @@ import mb.log.api.Logger;
 import mb.log.api.LoggerFactory;
 import mb.pie.api.ExecContext;
 import mb.pie.api.TaskDef;
-import mb.statix.multilang.LanguageId;
-import mb.statix.multilang.LanguageMetadata;
-import mb.statix.multilang.LanguageMetadataManager;
+import mb.statix.multilang.metadata.LanguageId;
+import mb.statix.multilang.metadata.LanguageMetadata;
+import mb.statix.multilang.metadata.LanguageMetadataManager;
 import mb.statix.multilang.MultiLang;
 import mb.statix.multilang.MultiLangScope;
 import mb.statix.multilang.spec.SpecBuilder;
