+++ core/statix.multilang/src/main/java/mb/statix/multilang/pie/SmlCheckTaskDef.java
@@ -14,12 +14,11 @@ import mb.pie.api.Task;
 import mb.pie.api.TaskDef;
 import mb.resource.ResourceKey;
 import mb.resource.hierarchical.ResourcePath;
-import mb.statix.multilang.AnalysisResults;
 import mb.statix.multilang.ConfigurationException;
-import mb.statix.multilang.ContextConfig;
-import mb.statix.multilang.LanguageId;
-import mb.statix.multilang.LanguageMetadata;
-import mb.statix.multilang.LanguageMetadataManager;
+import mb.statix.multilang.pie.config.ContextConfig;
+import mb.statix.multilang.metadata.LanguageId;
+import mb.statix.multilang.metadata.LanguageMetadata;
+import mb.statix.multilang.metadata.LanguageMetadataManager;
 import mb.statix.multilang.MultiLangAnalysisException;
 import mb.statix.multilang.pie.config.SmlBuildContextConfiguration;
 import mb.statix.multilang.utils.MessageUtils;
