+++ core/statix.multilang/src/main/java/mb/statix/multilang/pie/SmlPartialSolveFile.java
@@ -11,10 +11,8 @@ import mb.pie.api.Supplier;
 import mb.pie.api.TaskDef;
 import mb.resource.ResourceKey;
 import mb.statix.constraints.CUser;
-import mb.statix.multilang.FileResult;
-import mb.statix.multilang.ImmutableFileResult;
-import mb.statix.multilang.LanguageId;
-import mb.statix.multilang.LanguageMetadataManager;
+import mb.statix.multilang.metadata.LanguageId;
+import mb.statix.multilang.metadata.LanguageMetadataManager;
 import mb.statix.multilang.MultiLang;
 import mb.statix.multilang.MultiLangAnalysisException;
 import mb.statix.multilang.MultiLangScope;
