+++ core/statix.multilang/src/main/java/mb/statix/multilang/pie/SmlSolveProject.java
@@ -13,14 +13,9 @@ import mb.resource.hierarchical.ResourcePath;
 import mb.statix.constraints.CConj;
 import mb.statix.constraints.messages.IMessage;
 import mb.statix.constraints.messages.MessageUtil;
-import mb.statix.multilang.AnalysisResults;
-import mb.statix.multilang.FileKey;
-import mb.statix.multilang.FileResult;
-import mb.statix.multilang.ImmutableAnalysisResults;
-import mb.statix.multilang.ImmutableFileKey;
-import mb.statix.multilang.LanguageId;
-import mb.statix.multilang.LanguageMetadata;
-import mb.statix.multilang.LanguageMetadataManager;
+import mb.statix.multilang.metadata.LanguageId;
+import mb.statix.multilang.metadata.LanguageMetadata;
+import mb.statix.multilang.metadata.LanguageMetadataManager;
 import mb.statix.multilang.MultiLang;
 import mb.statix.multilang.MultiLangAnalysisException;
 import mb.statix.multilang.MultiLangScope;
@@ -37,7 +32,6 @@ import org.metaborg.util.task.NullCancel;
 import org.metaborg.util.task.NullProgress;
 
 import javax.inject.Inject;
-import java.io.File;
 import java.io.Serializable;
 import java.util.AbstractMap;
 import java.util.Collection;
