+++ core/statix.multilang/src/main/java/mb/statix/multilang/pie/config/MultiLangConfig.java
@@ -1,5 +1,7 @@
-package mb.statix.multilang;
+package mb.statix.multilang.pie.config;
 
+import mb.statix.multilang.metadata.ContextId;
+import mb.statix.multilang.metadata.LanguageId;
 import org.immutables.value.Value;
 
 import java.util.Map;
