+++ core/statix.multilang/src/main/java/mb/statix/multilang/pie/config/SmlBuildContextConfiguration.java
@@ -6,19 +6,14 @@ import mb.pie.api.ExecContext;
 import mb.pie.api.TaskDef;
 import mb.resource.hierarchical.ResourcePath;
 import mb.statix.multilang.ConfigurationException;
-import mb.statix.multilang.ContextConfig;
-import mb.statix.multilang.ContextDataManager;
-import mb.statix.multilang.ContextId;
-import mb.statix.multilang.ImmutableContextConfig;
-import mb.statix.multilang.LanguageId;
+import mb.statix.multilang.metadata.ContextDataManager;
+import mb.statix.multilang.metadata.ContextId;
+import mb.statix.multilang.metadata.LanguageId;
 import mb.statix.multilang.MultiLang;
-import mb.statix.multilang.MultiLangConfig;
 import mb.statix.multilang.MultiLangScope;
-import org.checkerframework.checker.nullness.qual.Nullable;
 
 import javax.inject.Inject;
 import java.io.Serializable;
-import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.Map;
 import java.util.Objects;
