+++ core/statix.multilang/src/main/java/mb/statix/multilang/pie/config/SmlReadConfigYaml.java
@@ -5,10 +5,8 @@ import mb.pie.api.ExecContext;
 import mb.pie.api.TaskDef;
 import mb.resource.ReadableResource;
 import mb.resource.hierarchical.ResourcePath;
-import mb.statix.multilang.ContextId;
-import mb.statix.multilang.ImmutableMultiLangConfig;
-import mb.statix.multilang.LanguageId;
-import mb.statix.multilang.MultiLangConfig;
+import mb.statix.multilang.metadata.ContextId;
+import mb.statix.multilang.metadata.LanguageId;
 import mb.statix.multilang.MultiLangScope;
 import org.yaml.snakeyaml.Yaml;
 import org.yaml.snakeyaml.constructor.Constructor;
